Compras - cambios proveerodes

+++ src/main/java/com/jhndagon/app/jwt/models/Rol.java
@@ -1,16 +1,12 @@
 package com.jhndagon.app.jwt.models;
 
 import java.io.Serializable;
-import java.util.List;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import javax.validation.constraints.NotNull;
 
@@ -25,7 +21,7 @@ public class Rol implements Serializable {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	@NotNull
 	@Column(unique= true, length=20)
 	private String nombre;
